Make templates first-class members of our monorepo (#6381)
This diff aims to make working on our templates a first-class part of
our monorepo in preparation for the starter kits work we've been talking
about.

There are a few big differences here:
- Templates are now a part of the yarn workspace. Their dependencies are
installed with the others
- They're typechecked alongside everything else in the repo
- They run against the in-repo versions of our packages instead of the
release versions
- There's a new release script that hopefully will make adding new
starter kits really easy
- They're linted by our monorepo lint tooling
- There's a new top-level `yarn dev-template <template name>` script to
make it easy to start work on a template.

This has resulted in a few side effects for the examples themselves:
- There are some config options needed only when in the monorepo. These
can be wrapped in special comments to exclude them from the published
versions
- There's a special declaration in package.json that describes how they
should be published, including removing certain things that we don't
want published
- The published examples no longer include linting
- I simplified the tsconfig setups of several of them. They now all use
a single tsconfig file, which fits better with our system.
- Simplifying tsconfig prompted some other changes - i switched the
cloudflare sync example over to use the vite cloudflare plugin instead
of separate vite/wrangler builds.
- I fixed some mistakes, type errors, and linting issues that enabling
these checks exposed

I also fixed up / deleted some unrelated cruft & mistakes that i think
come from vibe-coding. Working on this caused a few dependencies to get
updated when the lockfile was invalidated (prettier, typescript) so I've
upgraded the package.json files across the board to reflect the newer
versions.

There's some prior art in #6051 which aimed to integrate the templates
with our type checking while keeping them outside the rest of the
monorepo flow. This diff goes a fair bit further.

- [x] `other`

+++ packages/editor/src/lib/license/LicenseManager.test.ts
@@ -417,7 +417,7 @@ function importPrivateKey(pemContents: string) {
 	// base64 decode the string to get the binary data
 	const binaryDerString = atob(pemContents)
 	// convert from a binary string to an ArrayBuffer
-	const binaryDer = str2ab(binaryDerString) as Uint8Array
+	const binaryDer = str2ab(binaryDerString)
 
 	return crypto.subtle.importKey(
 		'pkcs8',
